initial

+++ backend/controllers/cart.controller.js
@@ -1,8 +1,10 @@
 const cartModel=require('../model/cart.model')
+const productModel = require('../model/product.model')
 
 const showCart=async(req,res)=>{
     try{
-        const cartItems= await cartModel.find({})
+        const userId=req.user.username
+        const cartItems= await cartModel.find({"userId" : userId})
         console.log(cartItems)
         if(cartItems.length>0){
             res.status(200).json(cartItems)
@@ -18,9 +20,24 @@ const showCart=async(req,res)=>{
 
 const addToCart= async (req,res)=>{
     try{
-        const product= req.body
-        await cartModel.create(product)
-        res.status(200).json(product)
+
+        const cartItemId= req.params.id;
+        const quantity=req.params.quantity
+        const userId=req.user.username;
+        //console.log(cartItemId,quantity,userId)
+        const product= await productModel.findOne({"productId" : cartItemId})
+        const productName=product.productName
+        const price=parseInt(product.price) * parseInt(quantity)
+        console.log(product.productName)
+
+        await cartModel.create({
+            "cartItemId" : cartItemId,
+            "userId" : userId,
+            "productName" : productName,
+            "quantity" : quantity,
+            "price" : price
+        })
+        res.status(200).json({"message" : "Item added to cart"})
         
     }
     catch(error){
@@ -31,7 +48,8 @@ const addToCart= async (req,res)=>{
 const deleteCartItem= async (req,res)=>{
     try{
         const cartItemId= req.params.id
-        const item= await cartModel.findOneAndDelete({"cartItemId" : cartItemId})
+        const userId=req.user.username
+        const item= await cartModel.findOneAndDelete({"cartItemId" : cartItemId,"userId" : userId})
         console.log(item)
         if(!item)
             res.status(404).json({"message" : "product not found"})
